Fixed race due to incorrect order of operations.
In RPC, client.write() would be called before client.insert(). In
certain conditions (that primarily showed up on Linux), this could
result in the result being available to the client before the result was
registered. This would then cause the read thread to purge the result,
as it would think that it was no longer needed.