Reverted change that wasn't relevant.
The state of the result should be modified in the client. It doesn't
make sense to do it in the result, as that would only be relevant if
multiple threads could read and write to the client. However the client
is not thread-safe, a separate client should be used for each thread.