Modified socket to only call notify on condition.
This is a small optimisation that should reduce the amount of
ping-ponging between threads. Essential on wait() the result that is
being waited for is registered before waiting on the condition. Then in
call(), the result being waited on is compared against the result being
waited for, if they are a match, then notify_one() is called.