Modified build to not require compiler script.
The compiler script was mainly used to reduce the overly verbose output
from SCons. This can now be done directly within SCons and the compiler
script is no longer necessary. Currently the only missing functionality
is changing the compiler depending on file extension - which is not
required, but would be ideal for Objective-C++ under Mac OS X.