Fixed bug in migration test on Mac OS X.
Modified migration test to wait after sending SIGTERM to ensure that the
process is shut down. This comes in to play if the test is run over
several iterations where the server process may not have shut down
before the next iteration is run.

Additionally a safe guard was added to ensure that the semaphore used to
ensure the server processes were ready would not be triggered until
after server construction in each process.