Removed assert() and replaced with check().
This is a more simple method of verifying a test condition. Instead of
stopping at one failure, it will continue on with the testing to collect
as much information as possible. Check() will also gather information
about the file and line the failure came from, so it can more easily be
traced, which is especially useful for longer tests.