Fixed IPC migration on Mac OS X.
Previously, this exploited the fact that sem_t*'s on Mac OS X were just
file descriptors. In 10.6, sendmsg() now verifies the descriptor type so
this it is no longer possible to pass them through sockets. Terry
Lambert aslo informed me that it was not wise to exploit this feature as
the kernel was not able to track the semaphore properly.