Scaled back canary to terminate server.
This is instead of the previous behaviour that would cause an exception
to be thrown on the client side. The reason for this is due to Mac OS's
inability to handle exceptions from a signal handler, thereby preventing
proper clean up of stack allocated variables. Rather than use longjmp(),
the choice was made to terminate the server, and keep behaviour
consistent on both Linux and Mac OS X.