Fixed graphical glitch in virtual_texture.
It turns out that the implicit derivative calculated in a texture2D() is
undefined in a branch. This would cause white lines to appear on the
border of textures when a fault caused vaddress() to lookup a higher
mipmap level. To fix this, derivatives should be precalculated
explicitly and use texture2DGrad().