Fix lua bindings. (Still has GC bug.)
There's still a bug that required disabling deletes in the GC callback
function. If the deletes are enabled, when a member function pointer is
collected, a segmentation violation occurs. The Lua GC jumps to the
callback, but for some reason that causes a segmentation violation, even
though the address seems to be valid.